+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"E:\\study\\React\\react-study\\src\\index.js":"1","E:\\study\\React\\react-study\\src\\App.js":"2","E:\\study\\React\\react-study\\src\\Components\\Header\\index.js":"3","E:\\study\\React\\react-study\\src\\Components\\Layout\\index.js":"4","E:\\study\\React\\react-study\\src\\Components\\Footer\\index.js":"5","E:\\study\\React\\react-study\\src\\Components\\PokemonCard\\index.js":"6"},{"size":219,"mtime":1611574726323,"results":"7","hashOfConfig":"8"},{"size":4584,"mtime":1611918497071,"results":"9","hashOfConfig":"8"},{"size":358,"mtime":1611758537177,"results":"10","hashOfConfig":"8"},{"size":842,"mtime":1611917590492,"results":"11","hashOfConfig":"8"},{"size":277,"mtime":1611816092955,"results":"12","hashOfConfig":"8"},{"size":1984,"mtime":1611920205815,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"ig3wvq",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\study\\React\\react-study\\src\\index.js",[],["27","28"],"E:\\study\\React\\react-study\\src\\App.js",[],"E:\\study\\React\\react-study\\src\\Components\\Header\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\Layout\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\Footer\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\PokemonCard\\index.js",[],{"ruleId":"29","replacedBy":"30"},{"ruleId":"31","replacedBy":"32"},"no-native-reassign",["33"],"no-negated-in-lhs",["34"],"no-global-assign","no-unsafe-negation"]
+[{"E:\\study\\React\\react-study\\src\\index.js":"1","E:\\study\\React\\react-study\\src\\App.js":"2","E:\\study\\React\\react-study\\src\\Components\\Header\\index.js":"3","E:\\study\\React\\react-study\\src\\Components\\Layout\\index.js":"4","E:\\study\\React\\react-study\\src\\Components\\Footer\\index.js":"5","E:\\study\\React\\react-study\\src\\Components\\PokemonCard\\index.js":"6","E:\\study\\React\\react-study\\src\\routes\\Home\\index.js":"7","E:\\study\\React\\react-study\\src\\routes\\Game\\index.js":"8","E:\\study\\React\\react-study\\src\\Components\\NavBar\\index.js":"9","E:\\study\\React\\react-study\\src\\Components\\Menu\\index.js":"10","E:\\study\\React\\react-study\\src\\Components\\MenuNavbar\\index.js":"11"},{"size":219,"mtime":1611574726323,"results":"12","hashOfConfig":"13"},{"size":482,"mtime":1612179507254,"results":"14","hashOfConfig":"13"},{"size":546,"mtime":1612179499682,"results":"15","hashOfConfig":"13"},{"size":869,"mtime":1612174956768,"results":"16","hashOfConfig":"13"},{"size":277,"mtime":1611816092955,"results":"17","hashOfConfig":"13"},{"size":1973,"mtime":1612174996600,"results":"18","hashOfConfig":"13"},{"size":5103,"mtime":1612181433717,"results":"19","hashOfConfig":"13"},{"size":99,"mtime":1612176152524,"results":"20","hashOfConfig":"13"},{"size":558,"mtime":1612180955200,"results":"21","hashOfConfig":"13"},{"size":671,"mtime":1612182397363,"results":"22","hashOfConfig":"13"},{"size":411,"mtime":1612181733959,"results":"23","hashOfConfig":"13"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},"ig3wvq",{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"26"},{"filePath":"41","messages":"42","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"43","usedDeprecatedRules":"26"},{"filePath":"44","messages":"45","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"46","messages":"47","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"E:\\study\\React\\react-study\\src\\index.js",[],["48","49"],"E:\\study\\React\\react-study\\src\\App.js",[],"E:\\study\\React\\react-study\\src\\Components\\Header\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\Layout\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\Footer\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\PokemonCard\\index.js",[],"E:\\study\\React\\react-study\\src\\routes\\Home\\index.js",[],"E:\\study\\React\\react-study\\src\\routes\\Game\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\NavBar\\index.js",["50"],"import styles from \"./style.module.css\";\r\nimport cn from \"classnames\";\r\n\r\nconst NavBar = ({ openMenu, toggleMenu }) => {\r\n  const onToogleMenu = () => {\r\n    toggleMenu && toggleMenu();\r\n  };\r\n  return (\r\n    <nav id={styles.navbar}>\r\n      <div className={styles.navWrapper}>\r\n        <p className={styles.brand}>LOGO</p>\r\n        <a\r\n          className={cn(styles.menuButton, { [styles.active]: openMenu })}\r\n          onClick={onToogleMenu}\r\n        >\r\n          <span></span>\r\n        </a>\r\n      </div>\r\n    </nav>\r\n  );\r\n};\r\n\r\nexport default NavBar;\r\n","E:\\study\\React\\react-study\\src\\Components\\Menu\\index.js",[],"E:\\study\\React\\react-study\\src\\Components\\MenuNavbar\\index.js",[],{"ruleId":"51","replacedBy":"52"},{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","severity":1,"message":"56","line":12,"column":9,"nodeType":"57","endLine":15,"endColumn":10},"no-native-reassign",["58"],"no-negated-in-lhs",["59"],"jsx-a11y/anchor-is-valid","The href attribute is required for an anchor to be keyboard accessible. Provide a valid, navigable address as the href value. If you cannot provide an href, but still need the element to resemble a link, use a button and change it with appropriate styles. Learn more: https://github.com/evcohen/eslint-plugin-jsx-a11y/blob/master/docs/rules/anchor-is-valid.md","JSXOpeningElement","no-global-assign","no-unsafe-negation"]
