+++ src/Components/PokemonCard/style.module.css
@@ -1,3 +1,10 @@
+.root {
+    width: 268px;
+    height: 388px;
+    margin: 16px;
+    perspective: 1000px;
+}
+
 .pokemonCard {
     --fire: #fddfdf;
     --grass: #defde0;
@@ -20,13 +27,6 @@
     transform-style: preserve-3d;
     transition: 0.6s 0.1s;
     transform: rotateY(180deg);
-}
-
-.pokemonCard {
-    width: 268px;
-    height: 388px;
-    margin: 16px;
-    perspective: 1000px;
 }
 
 /* TODO: This place where we should change :hover on .active */
