+++ src/routes/Game/index.js
@@ -1,0 +1,5 @@
+const GamePage = () => {
+  return <div>This is game page!</div>;
+};
+
+export default GamePage;
